% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{check_version}
\alias{check_version}
\title{Check if an updated version is available}
\usage{
check_version(
  gh_user = "gadenbuie",
  gh_repo = "regexplain",
  this_version = packageVersion("regexplain")
)
}
\arguments{
\item{gh_user}{GitHub user account}

\item{gh_repo}{GitHub repo name}

\item{this_version}{The currently installed version of the package}
}
\description{
I included this because it can be difficult to tell if your RStudio Addins
are up to date. I may add new features that you want but you won't hear about
the updates. This function checks if an update is available, using GitHub
tags. If an update is available, a modal dialog is shown when you start
the regexplain gadget. This only happens once per R session, though, so feel
free to ignore the message.
}
\keyword{internal}
