% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_ui.R
\name{generate_help_ui}
\alias{generate_help_ui}
\title{Generates Help Tab UI}
\usage{
generate_help_ui(cheatsheet_only = TRUE)
}
\arguments{
\item{cheatsheet_only}{If TRUE then returns just basic regex
explainer UI.}
}
\description{
Generates Help Tab UI
}
\keyword{internal}
