% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_gadget.R
\name{get_templates}
\alias{get_templates}
\title{Loads Regex Pattern Templates}
\usage{
get_templates()
}
\description{
Sourced from \href{https://projects.lukehaas.me/regexhub}{Regex Hub}
and available at \url{https://github.com/lukehaas/RegexHub}. Copyright
Luke Haas licensed under the MIT license available at
\url{https://github.com/lukehaas/RegexHub/commit/3ab87b5a4fd2817b42e2e45dcf040d4f0164ea37}.
}
\keyword{internal}
