% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_gadget.R
\name{regex_gadget}
\alias{regex_gadget}
\title{RegExplain gadget}
\usage{
regex_gadget(text = NULL, pattern = NULL, start_page = if
  (is.null(text)) "Text" else "RegEx")
}
\arguments{
\item{text}{Text to explore in gadget (editable using interface)}

\item{pattern}{Regular Expression to edit or visualize using RegExplain}

\item{start_page}{Open gadget to this tab, one of \code{"Text"}, \code{"RegEx"},
\code{"Output"}, or \code{"Help"}}
}
\description{
The function behind the RegExplain Selection and RegExplain File
addins. Opens the RegExplain gadget interface in an RStudio viewer
pane.
}
