% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplain-deprecated.R
\name{regexplain-deprecated}
\alias{regexplain-deprecated}
\alias{regex_gadget}
\title{Deprecated or renamed functions}
\usage{
regex_gadget(...)
}
\arguments{
\item{...}{Arguments passed to the new or renamed functions.}
}
\value{
The result of the new or renamed function.
}
\description{
These functions in regexplain have been deprecated or renamed.

Use \code{\link[=regexplain_gadget]{regexplain_gadget()}} instead of \code{regex_gadget()}.
}
\keyword{internal}
