% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_cheatsheet.R
\name{regexplain_cheatsheet}
\alias{regexplain_cheatsheet}
\title{Regex Cheatsheet Quick Reference}
\usage{
regexplain_cheatsheet()
}
\description{
The function behind the RegExplain Cheatsheet addin. Opens a summary of
regular expression syntax -- the RegExplain cheatsheet -- in an RStudio
viewer pane.
}
