% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{regexplain_file}
\alias{regexplain_file}
\title{Regexplain File Addin}
\usage{
regexplain_file()
}
\description{
See \link{regexplain_addin}. Opens file chooser to pick file, reads lines, returns
first \code{regexplain.addin.max_lines} (default 100).
}
\keyword{internal}
