% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{run_regex}
\alias{run_regex}
\title{Extract matched groups from regexp}
\usage{
run_regex(text, pattern, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, useBytes = FALSE, global = TRUE)
}
\arguments{
\item{text}{Text to search}

\item{pattern}{regexp}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}

\item{useBytes}{logical.  If \code{TRUE} the matching is done
    byte-by-byte rather than character-by-character.  See
    \sQuote{Details}.}

\item{global}{If \code{TRUE}, enables global pattern matching}
}
\description{
Extract matched groups from regexp
}
