% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{wrap_regex}
\alias{wrap_regex}
\title{Wraps capture groups in regex pattern in span tags to colorize with CSS}
\usage{
wrap_regex(pattern, escape = TRUE, exact = TRUE)
}
\arguments{
\item{pattern}{Regex pattern to look for}

\item{escape}{Escape HTML-related characters in \code{text}?}

\item{exact}{Should the regex pattern be displayed as entered by the user
into R console or source (default)? When \code{TRUE}, regex is displayed with
the double \code{\\\\} required for escaping backslashes in R. When \code{FALSE},
regex is displayed as interpreted by the regex engine (i.e. double \code{\\\\}
as a single \code{\\}).}
}
\description{
Wraps capture groups in regex pattern in span tags to colorize with CSS
}
\keyword{internal}
