% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{wrap_result}
\alias{wrap_result}
\title{Wrap matches in HTML span tags to colorize via CSS}
\usage{
wrap_result(x, escape = FALSE, exact = FALSE)
}
\arguments{
\item{x}{Individual list item in list returned by \code{\link[=run_regex]{run_regex()}}}

\item{escape}{Escape HTML-related characters in \code{text}?}

\item{exact}{Should the regex pattern be displayed as entered by the user
into R console or source (default)? When \code{TRUE}, regex is displayed with
the double \code{\\\\} required for escaping backslashes in R. When \code{FALSE},
regex is displayed as interpreted by the regex engine (i.e. double \code{\\\\}
as a single \code{\\}).}
}
\description{
Wrap matches in HTML span tags to colorize via CSS
}
\keyword{internal}
