% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_gadget.R
\name{regex_gadget}
\alias{regex_gadget}
\title{regexplain gadget}
\usage{
regex_gadget(text = NULL, start_page = if (is.null(text)) "Text" else
  "Regex")
}
\arguments{
\item{text}{Text to explore in gadget (editable using interface)}

\item{start_page}{Open gadget to this tab, one of `"Text"`, `"Regex"`,
`"Output"`, or `"Help"`}
}
\description{
regexplain gadget
}
