% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplain-package.R
\docType{package}
\name{regexplain-package}
\alias{regexplain}
\alias{regexplain-package}
\title{regexplain: Rstudio Addin to Explain, Test and Build Regular Expressions}
\description{
A set of RStudio Addins to help interactively test and build 
    regular expressions. Provides a Shiny gadget interface for interactively
    constructing the regular expression and viewing the results from common 
    string-searching functions. The gadget interface includes a helpful 
    regex syntax reference sheet and a library of common patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/regexplain}
  \item Report bugs at \url{https://github.com/gadenbuie/regexplain/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{g.adenbuie@gmail.com}

Other contributors:
\itemize{
  \item Winston Chang (Author of textInput and textAreaInput fragments from shiny) [contributor]
  \item Yihui Xie (Author of escape_html fragment from highr) [contributor]
  \item  RStudio (Copyright holder of included shiny fragments) [copyright holder]
}

}
\keyword{internal}
