% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modified_inputs.R
\name{textInputCode}
\alias{textInputCode}
\title{Modified Text Input}
\usage{
textInputCode(inputId, label, value = "", width = NULL,
  placeholder = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{Width of \code{shiny-input-container} div.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\description{
Standard \code{\link[shiny:textInput]{shiny::textInput()}} with additional \code{width} parameter, added code
font style for the input text and with \code{autocomplete}, \code{autocorrect},
\code{autocapitalize} and \code{spellcheck} set to \code{off} or \code{false}.
}
\seealso{
Other modified shiny inputs: \code{\link{textAreaInputAlt}}
}
\concept{modified shiny inputs}
