% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regex.R
\name{view_regex}
\alias{view_regex}
\title{View grouped regex results}
\usage{
view_regex(text, pattern, ..., render = TRUE, escape = render,
  knitr = FALSE, exact = escape)
}
\arguments{
\item{text}{Text to search}

\item{pattern}{Regex pattern to look for}

\item{...}{Passed to [run_regex]}

\item{render}{Render results to an HTML doc and open in RStudio viewer?}

\item{escape}{Escape HTML-related characters in `text`?}

\item{knitr}{Print into knitr doc? If `TRUE`, marks text as `asis_output` and
sets `render = FALSE` and `escape = TRUE`.}

\item{exact}{Should regex be displayed as entered by the user into R console
or source (default)? When `TRUE`, regex is displayed with the double `\\`
required for escaping backslashes in R. When `FALSE`, regex is displayed
as interpreted by the regex engine (i.e. double `\\` as a single `\`).}
}
\description{
View grouped regex results
}
